refactor: Consolidação do BytecodeGraph e Mapeamento de ExecutionMetadata (Spans) (#18)
* docs: remove excessive and redundant documentation in favor of streamlined runtime and architecture summaries

* refactor: remove serde dependencies and serialization modules from the galfus-image crate

* refactor: expand Constant enum with explicit bit-width integer and float variants and update IR builder logic

* refactor: rename galfus-image crate to galfus-bytecode and associated types for consistency

* refactor: modularize workspace state into source, semantic, and bytecode components

* feat: implement atomic transaction support for BytecodeGraph updates

* refactor: move module linking logic into bytecode crate and remove runtime module graph management

* refactor: update VirtualMachine to support multi-module execution via lifetime-bounded call frames and image resolution.

* refactor: simplify Runtime architecture, rename Bytecode structures, and implement stack trace formatting

* refactor: extract VM instruction execution into a public step method and make ExecutionStep public

* refactor: replace static bytecode linking with dynamic graph-based resolution in the runtime

* refactor: remove bytecode serialization, clarify in-memory-only architecture, and update documentation status

* refactor: rename Image types and fields to Bytecode terminology throughout the codebase

* refactor: implement versioned bytecode graph transactions with atomic validation and snapshot application

* feat: implement per-module global state and dependency-aware initialization order

* refactor: update Runtime to store BytecodeGraph and Providers for execution-scoped initialization

* refactor: rename PC to instruction_offset and implement source span resolution in panic stack traces

* refactor: introduce explicit ExportKind and ImportKind to support both functions and globals in bytecode modules

* refactor: wrap instructions and terminators in tuples to support optional source mapping metadata

+++ crates/galfus-frontend/src/parser/mod.rs
@@ -79,12 +79,12 @@ impl Parser {
 
     fn token_text(&self, token: &Token) -> &str {
         let span = token.span();
-        &self.source_text[span.start() as usize..span.end() as usize]
+        &self.source_text[span.start()..span.end()]
     }
 
     fn node_text(&self, node: NodeId) -> &str {
         let span = self.node_span(node);
-        &self.source_text[span.start() as usize..span.end() as usize]
+        &self.source_text[span.start()..span.end()]
     }
 
     pub fn finish(mut self) -> ParseResult {
